+++ public/app/Console/KhsCI/Repo/CacheCommand.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Console\Khsci\Repo;
+namespace App\Console\KhsCI\Repo;
 
 use Symfony\Component\Console\Command\Command;
 
-class StatusCommand extends Command
+class CacheCommand extends Command
 {
 }
