+++ public/app/Console/KhsCI/Repo/EnableCommand.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Console\Khsci\Repo;
+namespace App\Console\KhsCI\Repo;
 
 use Symfony\Component\Console\Command\Command;
 
-class ShowCommand extends Command
+class EnableCommand extends Command
 {
 }
