+++ public/app/Console/KhsCI/Repo/HistoryCommand.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Console\Khsci\Repo;
+namespace App\Console\KhsCI\Repo;
 
 use Symfony\Component\Console\Command\Command;
 
-class RequestsCommand extends Command
+class HistoryCommand extends Command
 {
 }
