+++ public/app/Console/KhsCI/Repo/ShowCommand.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Console\Khsci\Repo;
+namespace App\Console\KhsCI\Repo;
 
 use Symfony\Component\Console\Command\Command;
 
-class EncryptCommand extends Command
+class ShowCommand extends Command
 {
 }
