+++ public/app/Console/KhsCI/Repo/StatusCommand.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Console\Khsci;
+namespace App\Console\KhsCI\Repo;
 
 use Symfony\Component\Console\Command\Command;
 
-class OpenCommand extends Command
+class StatusCommand extends Command
 {
 }
