+++ public/public/demo/login/index.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Demo Login</title>
+</head>
+<body>
+
+</body>
+</html>
